Added prompting

+++ exportfoldertree.py
@@ -36,14 +36,13 @@ from sharedfunctions import getfolderid, callrestapi
 # get python version  
 version=int(str(sys.version_info[0]))
 
-
 # CHANGE THIS VARIABLE IF YOUR CLI IS IN A DIFFERENT LOCATION
 clidir='/opt/sas/viya/home/bin/'
-
 
 # get input parameters	
 parser = argparse.ArgumentParser(description="Export the complete Viya folder tree")
 parser.add_argument("-d","--directory", help="Directory for Export",required='True')
+parser.add_argument("-q","--quiet", help="Suppress the are you sure prompt.", action='store_true')
 args= parser.parse_args()
 basedir=args.directory
 
@@ -51,53 +50,65 @@ basedir=args.directory
 #now=time.localtime()
 #dirname="D_"+time.strftime('%Y-%m-%dT%H:%M:%S',now)
 #path=os.path.join(basedir,dirname)
-
 # I now just use the directory the user specifies
-path=basedir
-if not os.path.exists(path): os.makedirs(path)
-else: 
-    filelist=glob.glob(path+"/*.json")
-    for file in filelist:
-        os.remove(file)
 
-# retrieve root folders
-reqtype='get'
-reqval='/folders/rootFolders'
-resultdata=callrestapi(reqval,reqtype)
+# if the quiet mode flag is not passed then prompt to continue
+if not quitemode:
+	 if version  > 2:
+            areyousure=input("If the folder already exists any existing json files in it will be deleted. Continue? (Y)")
+        else:
+            areyousure=raw_input("If the folder already exists any existing json files in it will be deleted. Continue? (Y)") 
+else:
+	areyousure="Y"
 
 
-# loop root folders
-if 'items' in resultdata:
-  
-    total_items=resultdata['count']
-        
-    returned_items=len(resultdata['items'])
-        
-    if total_items == 0: print("Note: No items returned.")
-    else:
-     
-        # export each folder and download the package file to the directory
-		
-	for i in range(0,returned_items):   
+if areyousure.upper() =='Y':
+
+	path=basedir
+	if not os.path.exists(path): os.makedirs(path)
+	else: 
+		filelist=glob.glob(path+"/*.json")
+		for file in filelist:
+			os.remove(file)
+
+	# retrieve root folders
+	reqtype='get'
+	reqval='/folders/rootFolders'
+	resultdata=callrestapi(reqval,reqtype)
+
+
+	# loop root folders
+	if 'items' in resultdata:
+	
+		total_items=resultdata['count']
 			
-	     id=resultdata['items'][i]["id"]
-             package_name=str(uuid.uuid1())
-	     json_name=resultdata['items'][i]["name"].replace(" ","")+'_'+str(i)
+		returned_items=len(resultdata['items'])
+			
+		if total_items == 0: print("Note: No items returned.")
+		else:
+			# export each folder and download the package file to the directory
+			for i in range(0,returned_items):   
+				
+				id=resultdata['items'][i]["id"]
+					package_name=str(uuid.uuid1())
+				json_name=resultdata['items'][i]["name"].replace(" ","")+'_'+str(i)
+							
+				command=clidir+'sas-admin transfer export -u /folders/folders/'+id+' --name "'+package_name+'"'
+				print(command)     
+				subprocess.call(command, shell=True)
+
+				reqval='/transfer/packages?filter=eq(name,"'+package_name+'")'                
+				package_info=callrestapi(reqval,reqtype)
 					
-	     command=clidir+'sas-admin transfer export -u /folders/folders/'+id+' --name "'+package_name+'"'
-	     print(command)     
-	     subprocess.call(command, shell=True)
+				package_id=package_info['items'][0]['id']
+					
+				completefile=os.path.join(path,json_name+'.json')
+				command=clidir+'sas-admin transfer download --file '+completefile+' --id '+package_id    
+				print(command)
+				subprocess.call(command, shell=True)
 
-	     reqval='/transfer/packages?filter=eq(name,"'+package_name+'")'                
-	     package_info=callrestapi(reqval,reqtype)
-			
-	     package_id=package_info['items'][0]['id']
-			
-	     completefile=os.path.join(path,json_name+'.json')
-	     command=clidir+'sas-admin transfer download --file '+completefile+' --id '+package_id    
-	     print(command)
-	     subprocess.call(command, shell=True)
+	print("NOTE: Viya root folders exported to json files in "+path)
 
-print("NOTE: Viya root folders exported to json files in "+path) 
-         
+else: 
+	print"NOTE: Operation cancelled")
